% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_baum_welch.R
\name{baumWelchRecursion}
\alias{baumWelchRecursion}
\title{Apply Baum-Welch algorithm recursively}
\usage{
baumWelchRecursion(current_transition_matrix, observations, theta, p, b,
  constrained = FALSE, config_file = NULL)
}
\arguments{
\item{current_transition_matrix}{- Matrix, the current estiamtion of transition matrix}

\item{observations}{- Vector, vector of observations}

\item{theta}{- Numeric, exponential load parameter}

\item{p}{- Numeric, the stationary distribution parameter}

\item{b}{- Numeric, partition function size}
}
\value{
MLE for transition matrix
}
\description{
Apply BW single iteration
}
