% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_backward_vector.R
\name{computeBackwardVector}
\alias{computeBackwardVector}
\title{Compute forward vector for HMM}
\usage{
computeBackwardVector(hmm_object, observations)
}
\arguments{
\item{hmm}{- HiddenMarkovModel object}

\item{observation}{- vector of observations generated by the above class}
}
\value{
forward vector and loglikelihood
}
\description{
Computes the forward vector and loglikelihood for HMM semi parametric model
}
