% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_forward_vector.R
\name{computeForwardVector}
\alias{computeForwardVector}
\title{Compute forward vector for HMM}
\usage{
computeForwardVector(hmm_object, observations)
}
\arguments{
\item{observations}{- vector of observationss generated by the above class}

\item{hmm}{- HiddenMarkovModel object}
}
\value{
forward vector and loglikelihood
}
\description{
Computes the forward vector and loglikelihood for HMM semi parametric model
}
