% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_constrained_baum_welch_polynom.R
\name{computePolynomCoefficients}
\alias{computePolynomCoefficients}
\title{Computes polynom coefficients for S and T}
\usage{
computePolynomCoefficients(non_normalized_transition_matrix, p)
}
\arguments{
\item{non_normalized_transition_matrix}{- Matrix, count matrix from BW iteration}

\item{p}{- Numeric, the stationary distribution parameter}
}
\value{
List of polynom coefficients
}
\description{
Converts BW counts to polynom coefficients
}
