% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_baum_welch.R
\name{constrainedBaumWelch}
\alias{constrainedBaumWelch}
\title{Apply Baum-Welch algorithm}
\usage{
constrainedBaumWelch(hidden_markov_model_object, constrained = FALSE,
  number_of_iterations = NULL, config_file = NULL)
}
\arguments{
\item{constrained}{- Boolean, solve for general case or constrained?}

\item{number_of_iterations}{- Numeric, number of iteration}

\item{config_file}{- List, configuration file}

\item{hmm}{- HiddenMarkovModel object}
}
\value{
MLE for transition matrix
}
\description{
Computes the MLE for the transition matrix using constrained BW algorithm
}
