% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_baum_welch_solver.R
\name{constranedBaumWelchSolver}
\alias{constranedBaumWelchSolver}
\title{Solve constrained Baum-Welch equations}
\usage{
constranedBaumWelchSolver(bw_transition_matrix, observations, theta, p, b)
}
\arguments{
\item{bw_transition_matrix}{- Matrix, the result of BW}

\item{observations}{- Vector, vector of observations}

\item{theta}{- Numeric, exponential load parameter}

\item{p}{- Numeric, the stationary distribution parameter}

\item{b}{- Numeric, partition function size}
}
\value{
MLE for transition matrix
}
\description{
Solves the constrained BW equations
}
