% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_model_class.R
\name{createTransitionMatrix}
\alias{createTransitionMatrix}
\title{Create transition matrix for Semi-Parametric model}
\usage{
createTransitionMatrix(s, t, p = NULL, check_input = T,
  stationary_matrix = T)
}
\arguments{
\item{s}{- Numeric}

\item{t}{- Numeric}

\item{p}{- Numeric, default null, the value of the stationary vector}

\item{check_input}{- Logical, indicate if to check the caonstraints}

\item{stationary_matrix}{- Logical, indicate if the transition matrix is assumed to by stationary}
}
\value{
transition matrix
}
\description{
Create transition matrix for Semi-Parametric model
}
