% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_model_class.R
\name{createTransitionMatrixForHMM}
\alias{createTransitionMatrixForHMM}
\title{Create transition matrix function}
\usage{
createTransitionMatrixForHMM(theObject)
}
\arguments{
\item{s}{- float in range(0,1) indicating the probability of transition from 1/3->2}

\item{t}{- float in range(0,1) indicating the probability of transition from 1/3->1/3}
}
\value{
transition matrix
}
\description{
Create a 3*3 transition matrix for the semi parametric mode
}
