% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_observation_from_hmm.R
\name{generateHiddenSequance}
\alias{generateHiddenSequance}
\title{Generate samples of hidden states from HMM}
\usage{
generateHiddenSequance(n_obs, transition_matrix, delta)
}
\arguments{
\item{n_obs}{- Numeric}

\item{transition_matrix}{- Matrix, transition matrix}

\item{delta}{- Numeric, stationary vector}
}
\value{
vector of length \code{n_obs} of hidden states
}
\description{
Generate samples of hidden states from HMM
}
