% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_observation_from_hmm.R
\name{generateObservedSequance}
\alias{generateObservedSequance}
\title{Generate observation from HMM}
\usage{
generateObservedSequance(hidden_states, parameters_for_distribution,
  type = c("Normal", "Laplace"))
}
\arguments{
\item{hidden_states}{- Numeric, vector of hidden states}

\item{parameters_for_distribution}{- List, emission distribution parameters}

\item{type}{- Character, type of emission distribution}
}
\value{
list of hidden states and observations
}
\description{
Generate observation from HMM
}
