% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_model_class.R
\docType{class}
\name{HiddenMarkovModel-class}
\alias{HiddenMarkovModel-class}
\alias{hiddenMarkovModel}
\title{An S4 class to represent a HMM with exponential perturbations.}
\description{
An S4 class to represent a HMM with exponential perturbations.
}
\section{Slots}{

\describe{
\item{\code{sigma}}{numeric, the data standard deviation}

\item{\code{epsilon}}{numeric, the proportion of the mean wrt to std}

\item{\code{theta}}{numeric, exponential perturbation size}

\item{\code{p}}{numeric, stationary vector proportion}

\item{\code{s}}{numeric, transition probability to non-perturbed state}

\item{\code{t}}{numeric, transition probability between perturbed states}

\item{\code{n_obs}}{numeric, length of sample}

\item{\code{type}}{character, type of conditional distribution}

\item{\code{transition_matrix}}{matrix, HMM transition matrix}

\item{\code{observations}}{list, Observed data and Hidden data}
}}

