% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_observation_from_hmm.R
\name{sampleDataFromDoubleExponentialDistribution}
\alias{sampleDataFromDoubleExponentialDistribution}
\title{Sampling from A-Symmetric double Laplace}
\usage{
sampleDataFromDoubleExponentialDistribution(n, mu = 0)
}
\arguments{
\item{n}{- int, number of observations}

\item{mu}{- numeric, drift parameter}
}
\value{
x - n samples from A-symmetric laplace distribution
}
\description{
Sampling from A-Symmetric double Laplace
}
