% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_observation_from_hmm.R
\name{sampleDataFromShiftedNormalDistribution}
\alias{sampleDataFromShiftedNormalDistribution}
\title{Sample from Normal distribution}
\usage{
sampleDataFromShiftedNormalDistribution(n, theta, mu = 0, sigma2 = 1)
}
\arguments{
\item{n}{- Numeric, number of samples}

\item{theta}{- Numeric, perturbation parameter}

\item{mu}{- Numeric, location parameter}

\item{simga2}{- Numeric, variance parameter}
}
\value{
list of hidden states and observations
}
\description{
Sample from Normal distribution
}
