% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_constrained_baum_welch_polynom.R
\name{solveForS}
\alias{solveForS}
\title{Solves polynom for S}
\usage{
solveForS(polynom_coefficients)
}
\arguments{
\item{polynom_coefficients}{- List, polynom coefficients}
}
\value{
vector of solution for s
}
\description{
Compute solutions to quadratic equation for S
}
