% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_constrained_baum_welch_polynom.R
\name{solveForT}
\alias{solveForT}
\title{Solves polynom for T}
\usage{
solveForT(feasible_s_value, polynom_coefficients)
}
\arguments{
\item{feasible_s_value}{- List, feasible solutions for S}

\item{polynom_coefficients}{- List, polynom coefficients}
}
\value{
vector of solution for t
}
\description{
Compute solutions to linear equation for T
}
